Unify material reader
