+++ source/material_model/steinberger.cc
@@ -35,270 +35,6 @@ namespace aspect
   {
     namespace internal
     {
-
-      class MaterialLookup
-      {
-        public:
-          MaterialLookup(const std::string &filename,
-                         const bool interpol,
-                         const MPI_Comm &comm)
-          {
-
-            /* Initializing variables */
-            interpolation = interpol;
-            delta_press=-1.0;
-            min_press=-1.0;
-            delta_temp=-1.0;
-            min_temp=-1.0;
-            numtemp=0;
-            numpress=0;
-
-            std::string temp;
-            // Read data from disk and distribute among processes
-            std::istringstream in(Utilities::read_and_distribute_file_content(filename, comm));
-
-            getline(in, temp); // eat first line
-            getline(in, temp); // eat next line
-            getline(in, temp); // eat next line
-            getline(in, temp); // eat next line
-
-            in >> min_temp;
-            getline(in, temp);
-            in >> delta_temp;
-            getline(in, temp);
-            in >> numtemp;
-            getline(in, temp);
-            getline(in, temp);
-            in >> min_press;
-            min_press *= 1e5;  // conversion from [bar] to [Pa]
-            getline(in, temp);
-            in >> delta_press;
-            delta_press *= 1e5; // conversion from [bar] to [Pa]
-            getline(in, temp);
-            in >> numpress;
-            getline(in, temp);
-            getline(in, temp);
-            getline(in, temp);
-
-            Assert(min_temp >= 0.0, ExcMessage("Read in of Material header failed (mintemp)."));
-            Assert(delta_temp > 0, ExcMessage("Read in of Material header failed (delta_temp)."));
-            Assert(numtemp > 0, ExcMessage("Read in of Material header failed (numtemp)."));
-            Assert(min_press >= 0, ExcMessage("Read in of Material header failed (min_press)."));
-            Assert(delta_press > 0, ExcMessage("Read in of Material header failed (delta_press)."));
-            Assert(numpress > 0, ExcMessage("Read in of Material header failed (numpress)."));
-
-
-            max_temp = min_temp + (numtemp-1) * delta_temp;
-            max_press = min_press + (numpress-1) * delta_press;
-
-            density_values.reinit(numtemp,numpress);
-            thermal_expansivity_values.reinit(numtemp,numpress);
-            specific_heat_values.reinit(numtemp,numpress);
-            vp_values.reinit(numtemp,numpress);
-            vs_values.reinit(numtemp,numpress);
-            enthalpy_values.reinit(numtemp,numpress);
-
-            unsigned int i = 0;
-            while (!in.eof())
-              {
-                double temp1,temp2;
-                double rho,alpha,cp,vp,vs,h;
-                in >> temp1 >> temp2;
-                in >> rho;
-                if (in.fail())
-                  {
-                    in.clear();
-                    rho = density_values[(i-1)%numtemp][(i-1)/numtemp];
-                  }
-                in >> alpha;
-                if (in.fail())
-                  {
-                    in.clear();
-                    alpha = thermal_expansivity_values[(i-1)%numtemp][(i-1)/numtemp];
-                  }
-                in >> cp;
-                if (in.fail())
-                  {
-                    in.clear();
-                    cp = specific_heat_values[(i-1)%numtemp][(i-1)/numtemp];
-                  }
-                in >> vp;
-                if (in.fail())
-                  {
-                    in.clear();
-                    vp = vp_values[(i-1)%numtemp][(i-1)/numtemp];
-                  }
-                in >> vs;
-                if (in.fail())
-                  {
-                    in.clear();
-                    vs = vs_values[(i-1)%numtemp][(i-1)/numtemp];
-                  }
-                in >> h;
-                if (in.fail())
-                  {
-                    in.clear();
-                    h = enthalpy_values[(i-1)%numtemp][(i-1)/numtemp];
-                  }
-
-                getline(in, temp);
-                if (in.eof())
-                  break;
-
-                density_values[i%numtemp][i/numtemp]=rho;
-                thermal_expansivity_values[i%numtemp][i/numtemp]=alpha;
-                specific_heat_values[i%numtemp][i/numtemp]=cp;
-                vp_values[i%numtemp][i/numtemp]=vp;
-                vs_values[i%numtemp][i/numtemp]=vs;
-                enthalpy_values[i%numtemp][i/numtemp]=h;
-
-                i++;
-              }
-            Assert(i==numtemp*numpress, ExcMessage("Material table size not consistent with header."));
-
-          }
-
-          double
-          specific_heat(double temperature,
-                        double pressure) const
-          {
-            return value(temperature,pressure,specific_heat_values,interpolation);
-          }
-
-          double
-          density(double temperature,
-                  double pressure) const
-          {
-            return value(temperature,pressure,density_values,interpolation);
-          }
-
-          double
-          thermal_expansivity(const double temperature,
-                              const double pressure) const
-          {
-            return value(temperature,pressure,thermal_expansivity_values,interpolation);
-          }
-
-          double
-          seismic_Vp(const double temperature,
-                     const double pressure) const
-          {
-            return value(temperature,pressure,vp_values,false);
-          }
-
-          double
-          seismic_Vs(const double temperature,
-                     const double pressure) const
-          {
-            return value(temperature,pressure,vs_values,false);
-          }
-
-          double
-          dHdT (const double temperature,
-                const double pressure) const
-          {
-            const double h = value(temperature,pressure,enthalpy_values,interpolation);
-            const double dh = value(temperature+delta_temp,pressure,enthalpy_values,interpolation);
-            return (dh - h) / delta_temp;
-          }
-
-          double
-          dHdp (const double temperature,
-                const double pressure) const
-          {
-            const double h = value(temperature,pressure,enthalpy_values,interpolation);
-            const double dh = value(temperature,pressure+delta_press,enthalpy_values,interpolation);
-            return (dh - h) / delta_press;
-          }
-
-          double
-          dRhodp (const double temperature,
-                  const double pressure) const
-          {
-            const double rho = value(temperature,pressure,density_values,interpolation);
-            const double drho = value(temperature,pressure+delta_press,density_values,interpolation);
-            return (drho - rho) / delta_press;
-          }
-
-          double
-          value (const double temperature,
-                 const double pressure,
-                 const dealii::Table<2,
-                 double> &values,
-                 bool interpol) const
-          {
-            const double nT = get_nT(temperature);
-            const unsigned int inT = static_cast<unsigned int>(nT);
-
-            const double np = get_np(pressure);
-            const unsigned int inp = static_cast<unsigned int>(np);
-
-            Assert(inT<values.n_rows(), ExcMessage("Attempting to look up a temperature value with index greater than the number of rows."));
-            Assert(inp<values.n_cols(), ExcMessage("Attempting to look up a pressure value with index greater than the number of columns."));
-
-            if (!interpol)
-              return values[inT][inp];
-            else
-              {
-                // compute the coordinates of this point in the
-                // reference cell between the data points
-                const double xi = nT-inT;
-                const double eta = np-inp;
-
-                Assert ((0 <= xi) && (xi <= 1), ExcInternalError());
-                Assert ((0 <= eta) && (eta <= 1), ExcInternalError());
-
-                // use these coordinates for a bilinear interpolation
-                return ((1-xi)*(1-eta)*values[inT][inp] +
-                        xi    *(1-eta)*values[inT+1][inp] +
-                        (1-xi)*eta    *values[inT][inp+1] +
-                        xi    *eta    *values[inT+1][inp+1]);
-              }
-          }
-
-
-
-        private:
-
-
-          double get_nT(double temperature) const
-          {
-            temperature=std::max(min_temp, temperature);
-            temperature=std::min(temperature, max_temp-delta_temp);
-            Assert(temperature>=min_temp, ExcMessage("ASPECT found a temperature less than min_T."));
-            Assert(temperature<=max_temp, ExcMessage("ASPECT found a temperature greater than max_T."));
-            return (temperature-min_temp)/delta_temp;
-          }
-
-          double get_np(double pressure) const
-          {
-            pressure=std::max(min_press, pressure);
-            pressure=std::min(pressure, max_press-delta_press);
-            Assert(pressure>=min_press, ExcMessage("ASPECT found a pressure less than min_p."));
-            Assert(pressure<=max_press, ExcMessage("ASPECT found a pressure greater than max_p."));
-            return (pressure-min_press)/delta_press;
-          }
-
-
-          dealii::Table<2,double> density_values;
-          dealii::Table<2,double> thermal_expansivity_values;
-          dealii::Table<2,double> specific_heat_values;
-          dealii::Table<2,double> vp_values;
-          dealii::Table<2,double> vs_values;
-          dealii::Table<2,double> enthalpy_values;
-
-
-          double delta_press;
-          double min_press;
-          double max_press;
-          double delta_temp;
-          double min_temp;
-          double max_temp;
-          unsigned int numtemp;
-          unsigned int numpress;
-          bool interpolation;
-      };
-
       LateralViscosityLookup::LateralViscosityLookup(const std::string &filename,
                                                      const MPI_Comm &comm)
       {
@@ -394,8 +130,8 @@ namespace aspect
     Steinberger<dim>::initialize()
     {
       for (unsigned i = 0; i < material_file_names.size(); i++)
-        material_lookup.push_back(std_cxx11::shared_ptr<internal::MaterialLookup>
-                                  (new internal::MaterialLookup(data_directory+material_file_names[i],interpolation,this->get_mpi_communicator())));
+        material_lookup.push_back(std_cxx11::shared_ptr<Lookup::PerplexReader>
+                                  (new Lookup::PerplexReader(data_directory+material_file_names[i],interpolation,this->get_mpi_communicator())));
       lateral_viscosity_lookup.reset(new internal::LateralViscosityLookup(data_directory+lateral_viscosity_file_name,this->get_mpi_communicator()));
       radial_viscosity_lookup.reset(new internal::RadialViscosityLookup(data_directory+radial_viscosity_file_name,this->get_mpi_communicator()));
       avg_temp.resize(n_lateral_slices);
